requirements.txt
setuptools>=68.0.0
wheel>=0.41.0

Flask==3.1.2
gunicorn==23.0.0

# Pin scikit-learn & joblib for compatibility with Python 3.13
scikit-learn==1.7.2
joblib==1.5.2
python-dotenv==1.0.1

numpy
pandas
scipy
simple-salesforce
requests
requests-toolbelt
zeep
pyjwt
cryptography
